Changes to Remove CoClient
Added Logic to Remove CoClient so it now can remove Primary Contacts as well
If Account is Primary Contact, Selects a household account to become the new Primary Contact
If there are are no more accounts attached to a household, instead it will set the Household to Inactive.
Tested on Swagger. Works.